+++ game-frontend/src/app/components/display-quadrant/display-quadrant.component.ts
@@ -88,6 +88,8 @@ export class DisplayQuadrantComponent extends BaseComponent implements OnInit {
         await this._missionService.sendAttackMission(this.myPlanet, this.selectedPlanet, this.selectedUnits).toPromise();
       } else if (this.missionType === 'COUNTERATTACK') {
         await this._missionService.sendCounterattackMission(this.myPlanet, this.selectedPlanet, this.selectedUnits).toPromise();
+      } else if (this.missionType === 'CONQUEST') {
+        await this._missionService.sendConquestMission(this.myPlanet, this.selectedPlanet, this.selectedUnits).toPromise();
       } else {
         throw new ProgrammingError(`Unexpected mission type ${this.missionType}`);
       }