+++ game-frontend/src/app/services/mission.service.ts
@@ -32,6 +32,10 @@ export class MissionService extends GameBaseService {
     return this._sendMission('mission/counterattack', sourcePlanet, targetPlanet, involvedUnits);
   }
 
+  public sendConquestMission(sourcePlanet: PlanetPojo, targetPlanet: PlanetPojo, involvedUnits: SelectedUnit[]): Observable<void> {
+    return this._sendMission('mission/conquest', sourcePlanet, targetPlanet, involvedUnits);
+  }
+
   private _sendMission(url: string, sourcePlanet: PlanetPojo, targetPlanet: PlanetPojo, involvedUnits: SelectedUnit[]): Observable<void> {
     return this._doPostWithAuthorizationToGame<UnitMissionInformation>(
       url, {